Get pagination of back and next working correctly
For back to be available, we need to be at any page that is
not the first page. For next, we count how many records there
are in total that match the query conditions and check if
the requested page is outside the bounds of the last page.

This commit also includes the completed rendering of the
table